Report kernel session id for the new user session
Requires logging to be set to VERBOSE.
This allows matching up processes's session id with the SSH session. This is useful for example when many people log to
the same account with different public keys or IP addresses, in order to audit which process has been started by which
remote endpoint or public key holder.